added core functionality required to compute StackMapTables (which is still WIP)
We now support two types of Labels: those based on names (Symbols) and those based on (old!) PCs. This enables the trivial computation of a map which stores the mapping from old to new pcs.

Added “verificationTypeInfo” to (Domain)Value to make it easy to compute the type infos required when computing StackMapTables

LabeledCode is now created using “LabeledCode(<Code object>)” (instead of CODE.toLabeledCode(<…>)

Instead of calling toCodeAttributeBuilder on LabeledCode it is now sufficient to just call “result”

Signed-off-by: Michael Eichberg <mail@michael-eichberg.de>